requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
Django==4.2.6
django-anymail==10.1
django-registration-redux==2.12
django-shopping-cart==0.1
idna==3.4
Pillow==10.0.1
pip==23.2.1
pycparser==2.21
requests==2.31.0
setuptools==68.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wheel==0.41.2
